Add files via upload
With this simple program connects the TTGO T-BEAN with the U-Blok home setup program, u-Center that you can download from : 
https://www.u-blox.com/en/product/u-center.
You must connect the first serial time 115200, as this is the initial configuration of the USB port. If you change the GPS serial  settings in the u-Center you must change the setup settings to the chosen speed, without changing the serial speed and generate the code again.